+++ src/main/java/org/betonquest/betonquest/quest/condition/entity/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.condition Conditions}
+ * for entity conditions.
+ */
+package org.betonquest.betonquest.quest.condition.entity;